Changed the firmware to fix an issue where there was no way to continue a bout to the next period if the bout timer had not expired. This can happen in the case of team competitions, where the target score (5,10,15 etc) can be achieved before the bout timer has expired.
Now, if the timer is stopped using the OK key on the IR handset, or the timer
expires, then pressing the * key will move to the next period, resetting the
bout timer but preserving the current score.

Pressing * immediately again will restart the entire bout, setting the period
back to 1 and clearing all scores.

Pressing * immediately again will change box mode to Stopwatch.

Note that the period number is not displayed on the box itself - it is displayed
on the Android-based repeater.