Fixed bug where status bar and navigation bar would stay on-screen even during other displays.
This used to work on older Android versions, but I had a condition that
attempted to use a different way of hiding the bars for Android R (11) or
later. My old phone was running Android 9, and my new one is running
Android 13, so it seems that the 'different' way does not work, but was not
tested until I got the new phone.

I've changed it to use the older method of hiding the bars, and this still
works on Android 13.