Fixed bug where the demo display did not put the left hit light in the right place.
This only occurred at the very startup of the app - after that, it worked perfectly.

The issue was due to the use of a variable before it was initialised, in the onMeasure()
method of the HitLightView class.

A similar bug occurred in the onMeasure() method of the CardLightView class, and resulted
in the same problem - the card lights were displayed in the wrong place.

Rearranging the code in the onMeasure() methods to ensure that the variables
were initialised before use has fixed the bug.

Also, the copyright message was updated to 2025, and the document
was updated to state that the app has been run on Android 13.